rename Map to HotelsMap
