+++ project/src/components/hotels-map/hotels-map.jsx
