+++ project/src/components/room-page/room-page.jsx
@@ -11,7 +11,7 @@ import Review from '../review/review.jsx';
 import Comment from '../comment/comment.jsx';
 import PageHeader from '../page-header/page-header.jsx';
 import { Redirect, useParams } from 'react-router-dom';
-import Map from '../map/map.jsx';
+import HotelsMap from '../hotels-map/hotels-map.jsx';
 
 function getCurrentHotel(hotels, id) {
   return hotels.find((hotel) => hotel.id === Number(id));
@@ -110,7 +110,7 @@ export default function RoomPage({ hotels, reviews, user }) {
             </div>
           </div>
           <section className="property__map map">
-            <Map hotels={hotels} activeHotel={currentHotel} />
+            <HotelsMap hotels={hotels} activeHotel={currentHotel} />
           </section>
         </section>
         <div className="container">
